requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
dj-database-url==2.0.0
Django
django_heroku
django-cloudinary-storage==0.3.0
django-bootstrap-datepicker-plus==5.0.3
django-tinymce
django-bootstrap-v5==1.0.11
django-cors-headers
django-bootstrap5==23.1
django_storages
gunicorn
idna==3.4
Pillow==9.5.0
pydantic==1.10.7
python-dateutil
requests==2.29.0
Reportlab
sqlalchemy
six==1.16.0
soupsieve==2.4
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise
django[argon2]
django_cloudinary_storage
django-dotenv
